+++ app/services/theme.js
@@ -1,5 +1,4 @@
-import { getOwner } from '@ember/application';
-import { computed, get } from '@ember/object';
+import { computed } from '@ember/object';
 import { alias, empty, not, equal } from '@ember/object/computed';
 import Service, { inject as service } from '@ember/service';
 
@@ -56,14 +55,12 @@ export default Service.extend({
     },
 
     _onProviderLoad(provider) {
-        const locale = getOwner(this).factoryFor(`locale:${this.get('i18n.locale')}/translations`).class;
         this.set('provider', provider);
         this.set('reviewableStatusCounts', provider.get('reviewableStatusCounts'));
         this.get('i18n').addGlobals({
             provider: {
                 id: this.get('provider.id'),
                 name: this.get('provider.name'),
-                type: get(locale, `documentType.${this.get('provider.preprintWord')}`),
             },
         });
         return provider;